requirements.txt
flake8==7.1.0
isort==5.13.2
python-dotenv==1.0.1
python-telegram-bot==13.15
urllib3==1.26.6
requests==2.32.3